Refactor the Interpolate trait and add the Interpolator trait.
This commit represents 99% of the rework. From now on, implementing the
API requires to provide the various interpolation implementations. This
is actually a good thing, because people will now be able to either use
the `impl_Interpolate!` macro, which implements the interpolation in a
very “math” way (using std::ops::* traits and float literals), or by
providing their own.